+++ configs/freedom-k66f/src/k66_buttons.c
@@ -96,9 +96,9 @@ void board_button_initialize(void)
  * Name: board_buttons
  ****************************************************************************/
 
-uint8_t board_buttons(void)
+uint32_t board_buttons(void)
 {
-  uint8_t ret = 0;
+  uint32_t ret = 0;
 
   if (kinetis_gpioread(GPIO_SW2))
     {
